fix score

+++ Assets/Scripts/ProjectSystem/ScoreManager.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 
 namespace ProjectSystem {
     public static class ScoreManager {
-        public static float HiScore { get; private set; }
-        public static float PastScore { get; private set; } = 0;
-        public static float CurrentScore => PuzzleGameManager.Chicken_count * 59472 - Timer.GameTime * 5000;
+        public static int HiScore { get; private set; }
+        public static int PastScore { get; private set; } = 0;
+        public static int CurrentScore => (int) (PuzzleGameManager.Chicken_count * 59472 - Timer.GameTime * 5000);
         private static string key = "hiScore";
         
         public static void Save() {
@@ -17,7 +17,7 @@ namespace ProjectSystem {
         }
 
         static ScoreManager() {
-            HiScore = PlayerPrefs.GetFloat(key, 0);
+            HiScore = PlayerPrefs.GetInt(key, 0);
         }
     }
 }
